requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
firecrawl-py>=0.0.16
openai>=1.3.0
python-dotenv>=1.0.0
jsonschema>=4.20.0
httpx>=0.25.0

